#212: Extract card generation
