+++ administration/src/components/generation/GenerationController.tsx
@@ -6,17 +6,10 @@ import styled from "styled-components";
 import {add} from 'date-fns';
 import GenerationForm from "./GenerationForm";
 import {useApolloClient} from "@apollo/client";
-import generateCardActivateModel from "../../util/generateCardActivateModel";
-import {CardActivateModel} from "../../generated/compiled";
-import {addCard, addCardVariables} from "../../graphql/verification/__generated__/addCard";
-import {ADD_CARD} from "../../graphql/verification/mutations";
-import generateHashFromHashModel from "../../util/generateHashFromHashModel";
-import uint8ArrayToBase64 from "../../util/uint8ArrayToBase64";
 import {AppToaster} from "../AppToaster";
 import GenerationFinished from "./GenerationFinished";
-import {generatePdf} from "./PdfFactory";
-import {CardInput} from "../../../__generated__/globalTypes";
 import downloadDataUri from "./downloadDataUri";
+import generateCards from "./generateCards";
 
 let idCounter = 0;
 
@@ -46,37 +39,9 @@ const GenerationController = () => {
     const [mode, setMode] = useState(Mode.input);
     const confirm = async () => {
         try {
-            const region = 0; // TODO: Add correct region
-            const activateModels = cardCreationModels.map(model => {
-                const cardType = model.cardType === CardType.gold
-                    ? CardActivateModel.CardType.GOLD
-                    : CardActivateModel.CardType.STANDARD
-                return generateCardActivateModel(
-                    `${model.forename} ${model.surname}`,
-                    region,
-                    model.expirationDate,
-                    cardType)
-            })
             setMode(Mode.loading)
-            const cardInputs: CardInput[] = await Promise.all(
-                activateModels.map(async (model) => {
-                    const hashModel = await generateHashFromHashModel(model)
-                    return {
-                        expirationDate: model.expirationDate,
-                        hashModelBase64: uint8ArrayToBase64(hashModel),
-                        totpSecretBase64: uint8ArrayToBase64(model.totpSecret)
-                    }
-                }))
-            const results = await Promise.all(
-                cardInputs.map(card =>
-                    client.mutate<addCard, addCardVariables>({mutation: ADD_CARD, variables: {card}}))
-            )
-            const fail = results.find(result => result.errors || !result.data?.success)
-            if (fail) throw Error(JSON.stringify(fail))
-
-            const pdfDataUri = generatePdf(cardCreationModels)
+            const pdfDataUri = await generateCards(client, cardCreationModels)
             downloadDataUri(pdfDataUri, "ehrenamtskarten.pdf")
-
             setMode(Mode.finished)
         } catch (e) {
             console.error(e)
